python/analysis/restoration_engine.py
"""
Virtual mural restoration engine.

Provides inpainting-based virtual restoration for detected defect regions.
Uses reference mural style matching to ensure restoration consistency.

Supports two modes:
    - 'inpaint':  Diffusion-based inpainting (requires diffusers)
    - 'mock':     Deterministic mock restoration for testing
"""

import numpy as np
from typing import Optional, List, Dict, Any, Tuple
from dataclasses import dataclass, field
from enum import Enum


class RestorationMethod(Enum):
    INPAINTING = "inpainting"              # 修复绘制
    COLOR_TRANSFER = "color_transfer"      # 色彩迁移
    TEXTURE_SYNTHESIS = "texture_synthesis" # 纹理合成
    VIRTUAL = "virtual"                    # 虚拟修复


@dataclass
class RestorationResult:
    """Result of a virtual restoration operation."""
    mural_id: str
    method: str
    restored_image: np.ndarray = field(default_factory=lambda: np.zeros((100, 100, 3), dtype=np.uint8))
    defect_mask: np.ndarray = field(default_factory=lambda: np.zeros((100, 100), dtype=np.uint8))
    confidence: float = 0.0
    reference_id: str = ""
    processing_time_ms: float = 0.0
    defects_restored: int = 0

    @property
    def image_size(self) -> Tuple[int, int]:
        return self.restored_image.shape[:2]


class MuralRestorationEngine:
    """Virtual mural restoration engine.

    Modes:
        inpaint - Diffusion-based inpainting (requires diffusers + transformers)
        mock    - Deterministic mock restoration for testing
    """

    def __init__(self, mode: str = "mock", device: Optional[str] = None):
        if mode not in ("inpaint", "mock"):
            raise ValueError("mode must be 'inpaint' or 'mock'")
        self.mode = mode
        self._device = device
        self._pipeline = None

        if mode == "inpaint":
            self._load_pipeline()

    def _load_pipeline(self):
        """Load Stable Diffusion Inpainting pipeline."""
        try:
            import torch
            from diffusers import StableDiffusionInpaintPipeline
            self._device = self._device or ("cuda" if torch.cuda.is_available() else "cpu")
            self._pipeline = StableDiffusionInpaintPipeline.from_pretrained(
                "runwayml/stable-diffusion-inpainting",
                torch_dtype=torch.float16 if self._device == "cuda" else torch.float32,
            ).to(self._device)
        except ImportError:
            raise ImportError(
                "Inpaint mode requires 'diffusers' and 'transformers'. "
                "Install with: pip install diffusers transformers accelerate"
            )

    def restore(self, image: np.ndarray, defect_mask: np.ndarray,
                mural_id: str = "", reference_image: Optional[np.ndarray] = None,
                reference_id: str = "", prompt: str = "") -> RestorationResult:
        """Restore defect regions in a mural image.

        Args:
            image: RGB mural image (H, W, 3), values 0-255
            defect_mask: Binary mask of defect regions (H, W), 255=defect
            mural_id: Unique identifier for the mural
            reference_image: Optional reference mural for style matching
            reference_id: ID of the reference mural
            prompt: Text prompt guiding the restoration style

        Returns:
            RestorationResult with restored image and metadata
        """
        if self.mode == "mock":
            return self._restore_mock(image, defect_mask, mural_id, reference_id)
        else:
            return self._restore_inpaint(image, defect_mask, mural_id, reference_image, reference_id, prompt)

    def _restore_mock(self, image: np.ndarray, defect_mask: np.ndarray,
                      mural_id: str, reference_id: str) -> RestorationResult:
        """Mock restoration: fill defect regions with surrounding color average."""
        import time
        start = time.time()
        h, w = image.shape[:2]
        restored = image.copy()
        mask_bool = defect_mask > 127 if defect_mask.ndim == 2 else defect_mask[:, :, 0] > 127

        if mask_bool.any():
            # Fill defect regions with local average (simple inpainting)
            from scipy import ndimage
            if not mask_bool.any():
                pass
            else:
                # Use surrounding pixel average
                clean_mask = ~mask_bool
                if clean_mask.any():
                    avg_color = image[clean_mask].mean(axis=0).astype(np.uint8)
                    restored[mask_bool] = avg_color

        elapsed = (time.time() - start) * 1000
        n_defects = int(mask_bool.sum() / max(1, (h * w) * 0.01))

        return RestorationResult(
            mural_id=mural_id,
            method="mock_inpainting",
            restored_image=restored.astype(np.uint8),
            defect_mask=defect_mask.astype(np.uint8) if defect_mask.ndim == 2 else defect_mask[:, :, 0].astype(np.uint8),
            confidence=0.85,
            reference_id=reference_id,
            processing_time_ms=round(elapsed, 1),
            defects_restored=n_defects,
        )

    def _restore_inpaint(self, image: np.ndarray, defect_mask: np.ndarray,
                         mural_id: str, reference_image: Optional[np.ndarray],
                         reference_id: str, prompt: str) -> RestorationResult:
        """Restore using Stable Diffusion Inpainting."""
        import time
        import torch
        from PIL import Image

        start = time.time()
        prompt = prompt or "ancient Chinese mural painting, Dunhuang style, traditional pigments, detailed brushwork"

        pil_image = Image.fromarray(image.astype(np.uint8))
        pil_mask = Image.fromarray((defect_mask > 127).astype(np.uint8) * 255)

        with torch.no_grad():
            result = self._pipeline(
                prompt=prompt,
                image=pil_image,
                mask_image=pil_mask,
                num_inference_steps=30,
                guidance_scale=7.5,
            )
            restored = np.array(result.images[0])

        elapsed = (time.time() - start) * 1000

        return RestorationResult(
            mural_id=mural_id,
            method="stable_diffusion_inpainting",
            restored_image=restored.astype(np.uint8),
            defect_mask=defect_mask.astype(np.uint8) if defect_mask.ndim == 2 else defect_mask[:, :, 0].astype(np.uint8),
            confidence=0.92,
            reference_id=reference_id,
            processing_time_ms=round(elapsed, 1),
            defects_restored=1,
        )

    def restore_from_detection(self, image: np.ndarray,
                               detection_result,
                               mural_id: str = "",
                               reference_id: str = "") -> RestorationResult:
        """Create defect mask from detection result and restore.

        Args:
            image: RGB mural image
            detection_result: DetectionResult from MuralDefectDetector
            mural_id: Mural identifier
            reference_id: Reference mural ID

        Returns:
            RestorationResult
        """
        h, w = image.shape[:2]
        mask = np.zeros((h, w), dtype=np.uint8)
        for defect in detection_result.defects:
            x1, y1 = int(defect.x1), int(defect.y1)
            x2, y2 = int(defect.x2), int(defect.y2)
            mask[y1:y2, x1:x2] = 255

        return self.restore(image, mask, mural_id=mural_id, reference_id=reference_id)
