python/requirements.txt
"""
🏛️ Mural Guardian Python SDK requirements
"""
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
transformers>=4.40.0
ultralytics>=8.3.0
requests>=2.28.0
